+++ doc-examples/src/main/java/react4j/examples/arez/step2/Footer.java
@@ -7,7 +7,6 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
-import react4j.core.React;
 import react4j.core.ReactElement;
 import react4j.dom.events.MouseEventHandler;
 import react4j.dom.proptypes.html.AnchorProps;
@@ -35,7 +34,7 @@ class Footer
     final FilterMode filterMode = AppData.viewService.getFilterMode();
     return
       footer( new HtmlProps().className( "footer" ),
-              React.createElement( FooterTodoCount_.TYPE ),
+              FooterTodoCount.create(),
               ul( new HtmlProps().className( "filters" ),
                   li( a( new AnchorProps()
                            .className( FilterMode.ALL == filterMode ? "selected" : "" )