+++ processor/src/test/resources/expected/com/example/arez/ComponentWithAnnotatedParameterEventHandler_.java
@@ -3,6 +3,7 @@ package com.example.arez;
 import elemental2.core.JsObject;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import jsinterop.annotations.JsType;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
@@ -12,20 +13,38 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
+import react4j.core.React;
 import react4j.core.ReactConfig;
+import react4j.core.ReactElement;
+import react4j.core.ReactNode;
 
 @ArezComponent(
     name = "ComponentWithAnnotatedParameterEventHandler"
 )
 @Generated("react4j.processor.ReactProcessor")
 class ComponentWithAnnotatedParameterEventHandler_ extends ComponentWithAnnotatedParameterEventHandler {
-  public static final ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> TYPE = getConstructorFunction();
+  private static final ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> TYPE = getConstructorFunction();
 
   @Nonnull
   private final ComponentWithAnnotatedParameterEventHandler.CustomHandler2 _handleFoo = create_handleFoo();
 
   @Nonnull
   private final ComponentWithAnnotatedParameterEventHandler.CustomHandler _handleFoo2 = create_handleFoo2();
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create() {
+    return React.createElement( TYPE );
+  }
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props) {
+    return React.createElement( TYPE, props );
+  }
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+    return React.createElement( TYPE, props, child );
+  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> getConstructorFunction() {