+++ processor/src/test/resources/expected/com/example/event_handler/BasicHandlerComponent_.java
@@ -3,6 +3,7 @@ package com.example.event_handler;
 import elemental2.core.JsObject;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
 import react4j.core.BaseProps;
@@ -10,14 +11,32 @@ import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
 import react4j.core.Procedure;
+import react4j.core.React;
 import react4j.core.ReactConfig;
+import react4j.core.ReactElement;
+import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class BasicHandlerComponent_ extends BasicHandlerComponent {
-  public static final ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> TYPE = getConstructorFunction();
+  private static final ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> TYPE = getConstructorFunction();
 
   @Nonnull
   private final Procedure _handleFoo = create_handleFoo();
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create() {
+    return React.createElement( TYPE );
+  }
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props) {
+    return React.createElement( TYPE, props );
+  }
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+    return React.createElement( TYPE, props, child );
+  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> getConstructorFunction() {