+++ processor/src/test/resources/expected/com/example/render/RenderAsElementComponent_.java
@@ -8,12 +8,29 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
+import react4j.core.React;
 import react4j.core.ReactConfig;
+import react4j.core.ReactElement;
 import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class RenderAsElementComponent_ extends RenderAsElementComponent {
-  public static final ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> TYPE = getConstructorFunction();
+  private static final ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> TYPE = getConstructorFunction();
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create() {
+    return React.createElement( TYPE );
+  }
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props) {
+    return React.createElement( TYPE, props );
+  }
+
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+    return React.createElement( TYPE, props, child );
+  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseState, NativeReactComponent> getConstructorFunction() {