+++ todomvc/src/main/java/react4j/todomvc/TodoEntry.java
@@ -41,6 +41,12 @@ class TodoEntry
     }
   }
 
+  @Nonnull
+  static ReactElement<BaseProps, NativeReactComponent> create()
+  {
+    return _create();
+  }
+
   @Override
   protected void componentDidConstruct()
   {